seperated components and types
