+++ packages/frontend/src/App.tsx
@@ -1,11 +1,9 @@
-import React, { useEffect, useCallback, useState } from 'react';
+import { useEffect, useCallback, useState } from 'react';
 import {
   addEdge,
   Background,
   BackgroundVariant,
   Connection,
-  Controls,
-  MiniMap,
   ReactFlow,
   useEdgesState,
   useNodesState,
@@ -15,50 +13,10 @@ import { getEdges } from './api/edges.ts';
 
 import Loading from './components/Loading';
 import ThemeSwitcher from './components/ThemeSwitcher';
+import ControlsComp from './components/Controls';
 import { useTheme } from 'next-themes';
-
-export interface NodeData {
-  id: string;
-  label: string;
-  position: { x: number; y: number };
-  type?: string;
-  selected?: boolean;
-  isConnectable?: boolean;
-  zIndex?: number;
-  positionAbsoluteX?: number;
-  positionAbsoluteY?: number;
-  dragging?: boolean;
-  targetPosition?: string;
-  sourcePosition?: string;
-}
-
-export interface EdgeData {
-  id: string;
-  animated: boolean;
-  data: string | string[];
-  style: React.CSSProperties;
-  selected: boolean;
-  source: string;
-  target: string;
-  sourceHandleId?: string | null;
-  targetHandleId?: string | null;
-  interactionWidth: number;
-  sourceX: number;
-  sourceY: number;
-  targetX: number;
-  targetY: number;
-  sourcePosition: string;
-  targetPosition: string;
-  label?: string | React.ReactNode;
-  labelStyle?: React.CSSProperties;
-  labelShowBg?: boolean;
-  labelBgStyle?: React.CSSProperties;
-  labelBgPadding?: [number, number];
-  labelBgBorderRadius?: number;
-  markerStart?: string;
-  markerEnd?: string;
-  pathOptions?: string;
-}
+import {EdgeData, NodeData} from "./utils/types.ts";
+import MinimapComp from "./components/Minimap.tsx";
 
 export default function App() {
   const [nodes, setNodes, onNodesChange] = useNodesState([]);
@@ -133,8 +91,8 @@ export default function App() {
         minZoom={0.1}
         proOptions={{ hideAttribution: true }}
       >
-        <Controls />
-        <MiniMap />
+          <ControlsComp/>
+          <MinimapComp/>
         <Background variant={BackgroundVariant.Dots} gap={12} size={1} />
       </ReactFlow>
     </div>
