+++ packages/frontend/src/components/Controls.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import {Controls} from "@xyflow/react";
+
+const ControlsComp: React.FC = () => {
+    return (
+          <Controls />
+
+    );
+}
+export default ControlsComp;
