+++ packages/frontend/src/components/Minimap.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import {MiniMap} from "@xyflow/react";
+
+const MinimapComp: React.FC = () => {
+    return (
+              <MiniMap />
+
+    );
+}
+export default MinimapComp;
